+++ Models/Developer.cs
@@ -6,7 +6,7 @@ namespace SalaryCalculator.Models
         public LeadDeveloper Manager { get; set; }
         object IManaged.Manager { get; set; }
 
-        public Developer(int id, double basicSalary): base (id, basicSalary)
+        public Developer(int id, decimal basicSalary): base (id, basicSalary)
         {
         }
     }
