+++ SalaryCalculator.cs
@@ -6,9 +6,9 @@ namespace SalaryCalculator
 {
     public class SalaryCalculator
     {
-        public double GetTotalSalaries(Employee[] employees)
+        public decimal GetTotalSalaries(Employee[] employees)
         {
-            double total = 0;
+            decimal total = 0;
             foreach (var emp in employees)
             {
                 total += emp.GetTotalSalary();
